requirements.txt
aiohttp~=3.8.3
aiosqlite~=0.17.0
discord.py
